pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "retail-sales-pipeline"
version = "0.1.0"
description = "Retail sales medallion pipeline utilities"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "boto3>=1.34.0,<2.0.0",
    "botocore>=1.34.0,<2.0.0",
    "duckdb>=1.1.0,<2.0.0",
    "pandas>=2.2.0,<3.0.0",
    "pyarrow>=18.0.0,<20.0.0",
    "python-dotenv>=1.0.1,<2.0.0",
]

[project.scripts]
retail-sales-pipeline = "retail_sales_pipeline.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]